Add files via upload
implements APF on the FHN model, specifically trying to recover [V, w] from a noisy measurement of V. The 1d observation and 2d latent  space has an error (the ambiguity is on line 104). How is the mean of g_int_func, the gaussian representing g(y_n | x_n) defined?