Add files via upload
This demo shows how to run the APF code on the FHN model. The output of the function is averaged and smoothed.

Note that the ode solver in simFHN.py is non deterministic and sometimes generates null output. There may also be numerical issues as the result of all particles in a given iteration going to zero.

To do: fix the shifting of the signal for I_ext (the input current magnitude)